pyproject.toml
[project]
name = "watchcat"
version = "0.1.0"
description = "not a dog, but a cat that watches stocks"
authors = [
    {name = "kim.hyunggi",email = "kim.hyunggi@lumir.space"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "rich (>=14.0.0,<15.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "influxdb-client (>=1.49.0,<2.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
